ImageWithShimmer => ImageWithLoading
